Adds new tool for starting a workflow (#5)
This adds a new tool that will start a workflow on Itential Server.  The
`start_workflow(...)` tool accepts a single required argument `name`
which is the name of the workflow to start.  It also accepts a number of
optional arguments for configuring a description, setting input
variables and/or setting a list of groups that have access.