Fixes issue coericing env var to proper type (#16)
When using environment variables to configure the MCP server, the
applicaiton would try to pass them as string values which is incorrect.
The correct behavior is to corece them to the proper type.  This change
will now coerce the env variables for the MCP server to the same type as
the default.

fixes #14