Add new tool to retreive jobs from operations manager
Operations Manager can be used to retrieve all jobs from Itential
Platform.  By default this includes all jobs in any state.  This update
adds a new tool to get all jobs from the server.  The function provides
an optional keyword argument to filter the returned list to a subset of
jobs based on the job status.