Fixes issue where call to devices tool doesn't return
When a device is reachable through multiple adapters, the API call will
only return one of the device entries.  In this situation, the code
needs to check for the number of unique devices instead of the total
device count otherwise the function is stuck in a loop.  This is now
fixed with this patch.