Update the inspect logic to avoid loading imported functions (#41)
The server will introspect the tools folder and load all public
functions as a MCP tool.  There is a bug in the original code where by
it will attempt to load any callable function even if it is imported by
the tooling.   This change fixes that problem and now only loads
functions defined in the module.