Adds new tools for working with command templates
This adds four new tools to the server, all designed to work with the
Command Templates feature in Itential Platform.  The first tool added is
`get_command_templates(...)` which will retrieve a list of the available
command templates.  The second tool is `describe_command_template(...)`
which can be used to get details about a specific command template.  The
third tool is `run_command_template(...)` which will kick off the
command template run. The fourth tool is `run_command(...)` which can be
used to run a command against one or more devices.

Note the commands for getting and running command templates are project
aware.  When calling the `describe_command_template(...)` or
`run_command_template(...)` tool, if the desired command template is in
a project, the project name must also be provided otherwise the tool
will only consider command templates in the global space.