Adds a simple in-memory key value store for caching
This adds a new thread safe, in memory key value store that can be used
for caching return values from API calls.   An instance of the Cache
object is added to the Context to allow tools to cache return values in
order to save on making duplicate API calls to retrieve data.  The use
of the cache object is up to each tool