Updates the workflows module (#44)
This refactors the workflows module to use the internal function from
the accounts module.  The lookup function in workflows has been replaced
by the one in accounts.  This commit also refactors all public function
arguments so they are fully annotated using the Python typing module
along with pydantic.  This will help to provide more clarity on the
arguments for a give function.