Adds configuration file and tags to the application (#48)
This feature adds two signifcant changes to the MCP server.  First it
adds a new configuration file option when starting the sever.  To use a
configuration file, add the `--config` command line option and point it
to the desired configuration file.  There is an annotated configuration
file example avaliable at `docs/mcp.conf.example`

This also adds support for include and exclude tags to the application.
Include and exclude tags can be used to explictly include or exclude
tools when the server starts.  This allows for different tools to be
available and map to different authentications to the server