Refactor creating MCP server instance (#56)
This refactors how the FastMCP server instance is created.  Instead of
the server being created in `run()` it is now refactored into a new
function called `new(...)`.  The function will create and return an
instance of the FastMCP server.  This change is necessary to build a
better tool testing framework.