Adds new tool for creating LCM resources
This adds a new tool that will create a Lifecycle Manager resource on
the server.  The tool, `create_resource()`, requires a name and JSON
schema.  It will check the server if a resource model by the same name
exists and will error if it does.  If it doesn't exist, the resource
model will be created.