New tool for working with device groups (#60)
This adds two new tools to the server for working with device groups in
Configuration Manager.  The first tool `get_device_groups` will return a
list of device groups where each element in the list represents a unique
device group configured on the server.  The second tool
`create_device_group` will create a new device group on the server.