Adds new tools for working with applications
This adds new tools to the server that provide operations for working
with applications running on Itential Platform.  The following new tools
are added with this change:

- `get_applications`
- `start_application`
- `stop_application`
- `restart_application`

The `get_applications` tool will return a list of applications that are
configured to run on Itential Platform along with their current version
and operational state.

The `start_application` tool can be used to transition an application
from the `STOPPED` state to the `RUNNING` state.

The `stop_application` tool can be used to transition an application
from the `RUNNING` state to the `STOPPED` state

The `restart_application` will restart a running application.