Adds new tools for working with adapters (#66)
This adds new tools for working with the operational state of adapters
on the server.  It adds the following new tools:

- `get_adapters`
- `start_adapter`
- `stop_adapter`
- `restart_adapter`

The `get_adapter` tool will retreive a list of all configured adapters
from the server.

The `start_adapter` will attempt to put a stopped adapter into a running
state.

The `stop_adapter` will attempt to put a running adapter into a stopped
state

The `restart_adapter` will restart a running adapter.

Note this PR cannot be merged before #65