Containerfile updates (#104)
* updated server.py docstrings and instructions

* let's check this one out

* test cases updated

* updated tools.md with a more comprehensive tools reference list

* Added run command to the Containerfile ENTRYPOINT

* Adding new command for calling tools (#101)

This adds a new command to the application that will call a tool
directly.  To call a tool use `itential-mcp call <tool> [--params
<params>]`  This command is useful for testing the tool calls agaisnt a
server without having to have an entire end to end system configured.

* Globalize server instructions as a variable (#105)

This change moves the server instructions to a globalized variable
instead of passing the string directly as a keywork parameter.  This
will aid in updating the instructions later if needed.  All test cases
have been updated to ensure passing.

* updated server.py docstrings and instructions

Rebased server.py with latest devel

* let's check this one out

* test cases updated

* accepted changes to server.py

* Finally!!!

---------

Co-authored-by: joksan.flores <joksan.flores@itential.com>
Co-authored-by: Peter Sprygada <privateip@users.noreply.github.com>